requirements.txt
flask==3.1.1
openai==1.95.0
python-dotenv==1.1.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
httplib2==0.22.0
line-bot-sdk==3.17.1
